fixed some values - bounce was starting at too low a coordinate

+++ Vector_Utils.c
@@ -1,5 +1,5 @@
 #include "Vector_Utils.h"
-vector CreateVector(int x, int y) {
+vector CreateVector(float x, float y) {
     vector v;
     v.x = x;
     v.y = y;
